e2e: fix flaky annotation checks
In the annotation tests, we check the current and last applied
annotation. It's possible the last change made to the resource included
removing or adding annotations. This will make it so we cannot
definitively have a set of annotations to compare against across the
current and last applied annotation. In practice the only annotation
that we expect to be added or removed in e2e tests is the cluster-name.
So, we can just avoid checking that particular annotation and avoid
flakiness for these tests.

Bug: 131429115
Change-Id: I3f04c3b5882f4c0efa09a67c99edebf535f23dc7