Add more logs to syllogi test
Typically, when our Syllogi test fails it fails with "git-importer did
not come up"

Which is not very useful. Getting a little bit of insight into the
operator/importer logs will likely be useful to diagnose the nature of
the issue.

This is especially useful because getting at a running test cluster is
quite tricky.

Note that this change is already running successfuly in Syllogi CI.

Change-Id: I2d1faa8ebc366c9e3c1229d0476c3c67da57e118