vet: fix infinite loop getting clusterconfigs
There was a bug introduced by changes to Parser where we would get stuck
in an infinite loop trying to look up ClusterConfigs on the cluster.
This lookup should have happened after a check seeing if Nomos is
installed on the cluster.

This fixes it so that check happens before doing the lookup in nomos
vet.
This change includes some incidental refactoring.

Fixes: 133244761
Change-Id: I2d35846ce0e87fa4c3a89543ae111927ef4ceb7a