syncer: don't sync status field
We already disallow configs with a status field in importer. Because of
the way we do serialization, resources in clusterconfigs and
namespaceconfigs could still contain empty status fields. This makes it
so we'll ignore it when computing the diff between actual and declared
resources on the cluster and not get into a fight with other controllers
updating the status field.

Fixes: 133430917
Fixes: 133321608
Change-Id: Ibe4c2a39033e148ed3a0b0a7f961947bba5ed85c