importer: disallow configs with status field
This ended up more difficult than expected. When we convert the
runtime.Unstructured objects we parse from yaml to the typed version,
some fields are populated with empty data. It makes it so we aren't able
to determine if a user left the field empty or just added empty data to
it. If we wanted to handle all these cases, we would need to have a
check for each native k8s type (e.g. CRD and namespace populate the
empty status subresource differently).

The easier option is to just use the runtime.Unstructured, which
directly matches the parsed yaml. We can check that a status subresource
was not specified at all. This required having an optional
runtime.Unstructured in ast.FileObject. We generate this when it isn't
specified.

Long term, we should be storing the unstructured version of the object
in ast.FileObject and convert to the typed version on an as-needed
basis.

Fixes: 133431580
Change-Id: I4d8492667c80e6eae872f85102e410aa61152652