convert importer to a controller
This is the first phase of converting git-importer into a proper
controller. This change makes it so we are using the controller-runtime
libraries for creating a controller and adding watches to it. This is
only a refactor at this point and there are no added features.

In subsequent changes, the old clients and informers will be replaced
with controller clients and caches. After that, importer will reconcile
namespaceconfigs and clusterconfigs to match repo state. This will
guarantee out of band changes to these CRs get reverted by this controller.

Bug: 118385500
Change-Id: Id41a4dde6d2b94bed0e8cf190c198bf1ff4ce664