Put RootPath in ParseOpts
The lifetime of the g-root path is exactly the same as all other
configuration for the parser. Modifying the ConfigManagement is the only
way to change the directory to read from, so the same Parser will never
be used with a diffrent g-root.

Refactoring to make moving ClusterName determination out of Parser
easier.

Bug: 135535529
Change-Id: I358bcaa7c61a8e6dbff84cf89a3a82cec0e18ef0