Factor out parsing from reading from fileystem
Break Parse into two stages - reading from the filesystem into a DRY
format, and hydrating/validating the configuration.

In the future this will allow testing to not rely on literally writing
to the filesystem. In the near-term it means nomos hydrate will only
need to read from the filesystem once rather than once-per-cluster.

This CL turns parsing into a four-step process, where each stage can be
performed and tested individually:
1) Read CRDs
2) Read objects from filesystem
3) Hydrate/validate configuration
4) Output hydrated configuration

Change-Id: Id36e497edb1a4981f25300e34d87ee27621b9f6d