Start parser_test refactoring
Most of the tests in parser_test.go don't literally need reading from a
filesystem in order to work or test the behavior they're looking for.

This CL begins the transition of parser_test.go to the new libraries,
which is now possible as the Parser's function of reading from the
filesystem has been decoupled from actually vetting and hydrating the
config hierarchy.

FlatRoot has been moved to the ast package. This is because
otherwise using it in TreeBuilder introduces a circular dependency.
BuildTree has been refactored to allow tests to just specify their list
of objects and get a FlatRoot back rather than having to construct it
themselves. Since this logic would otherwise be duplicated, it has been
extracted from its existing location in buildTree.

An additional condition has been added to hasStatusField in
DisallowedFieldsValidator. This is due to a quirk of transitioning
between standard and Unstructured object forms - some types (such as
Namespace) automatically create an empty .status field, triggering this
vet error when it shouldn't. It does mean it is allowed for people to
specify an empty status field now, but specifying such a thing would
have no effect if synced.

HydrateTestCase is a new set of helped methods and types for refactoring
parser_test. New tests are about 1/4th the size of older tests.

Tests that either literally need to read from a filesystem or would be
impractical to refactor to not require reading from the filesystem are
now in their own file: parser_read_files_test.

Change-Id: I38d858db9b3be36dabd13a9bcbb98e5d7b12e170