importer: fix genericresources comparisons
Previously comparing namespaceconfigs and clusterconfigs was broken. The
problem is the cluster version had the RawExtension.Raw field set and
the generated version had the RawExtension.Object field set. Comparison
would always return not equal.

The fix is to decode the GenericResources into a common representation
for comparison. Being able to decode serialized resources, requires
having said resources registered in the scheme. The Decoder needed to be
updated to update it's underlying scheme and the reconcile loop needed
to be updated update the scheme with resources included in Syncs.

This will stil make it so namespaceconfigs always get updated with the
latest token, regardless of whether the underlying resources have
changed. In a subsequent commit, this will be fixed along with the
relevant e2e tests.

Bug: 135766013
Change-Id: I599c75ec562273527332de6f097337c71d687142