importer: don't update namespaceconfig if only token changes
This makes it so the token isn't comsidered for equality comparisions in
importer. Previously, it would update NamespaceConfigs if the token
field of the one on the cluster didn't match the importer-generated one.

This fixes e2e tests that were relying on .spec.Token always being updated
for every NamespaceConfig, for every commit.

Fixes: 135766013
Change-Id: I27e82e3d1fafd78d8ff822806f2f107b64c0aa1c