Extract out getting the scope of a GroupKind
The discovery.APIInfo class is doing far more than it needs to in order
to do its two jobs. The two things it needs to know are:
1) Is the GroupKind namespaced?
2) Is the GroupVersionKind synced?

Ideally these would be two difference classes as these concerns are used
separately. This refactoring just removes the unnecessary work APIInfo
does, as well as creating the Scoper interface to highlight where a
simpler class can be extracted out.

Add test for previously untested behavior used when restarting syncer.

Change-Id: I183b17207db39015207208b949cab46a241d9786