Make OutputVisitor add metadata.namespace
AllConfigs is not the final output of what is applied to the cluster.
One of these differences is that metadata.namespace is not written to
objects in AllConfigs.

Since OutputVisitor knows which Namespace each object belongs in, it
already has the information necessary to write metadata.namespace for
each object.

This means callers of Parser don't have to worry about filling in this
information.

Change-Id: I63f33fc56f2b61546bcf8cad91d2bdf340c0c5e0