Allow missing ConfigManagement CRD
Previous CL was submitted while still WIP.

Tests currently fail if ConfigManagement CRD is not defined.

Delete CRD instead of ConfigManagement directly. If deletion succeeds we
know no ConfigManagements exist. If it times out, attempt to delete the
one we know of by removing the operator finalizer.

Does not handle the case of rogue ConfigManagements applied to the
cluster which have unactivated finalizers, but that should be an edge
case.

Also fix warnings about variable usage.

Change-Id: I1ca92fa033e6fa7b8f7347b48b6b0a329a09727e