Fix a race condition between syncer and importer.
The syncer and importer have a race condition where Sync update may get
processed after a namespace or the cluster is reconciled.  This leads to
the issue that resources inside of the ClusterConfig or NamespaceConfig
may not be applied to the API server, but the status for the
ClusterConfig or NamespaceConfig will still report that the resource has
been reconciled successfully.

Bug: 137871168
Change-Id: Ia31f7f8771a4bffc42da86fc634c02a89048a2e6