Add 2 domains to binary/operator ACL.
External training sessions need to be able to retrieve the nomos binary
and the operator to learn how to use Nomos.  In b/139345284 we've done
this manually for the ACM 1.0.0 version, but we don't want those labs to
break when we bless the next release.  In addition to adding access for
these domains, I've improved the explanataion and added a TODO so that
we remember to clean this up when we actually have a license enforcement
mechanism in the product.

Bug: 139345284
Change-Id: Id3c04a8a0d79ae67a16862cd5ec8d05071881f90