Add basic vet validation to RawParser
Add three types of validation to RawParser
1) Duplicate group/kind/namespace/name detection
2) Forbid ConfigManagement kinds
3) Forbid config-management-system Namespace

Until we update the Group of the operator, it isn't covered by this
implementation. It feels safe to not check for that since we're going to
change it before release.

Write basic validation and test framework for nonhierarchical logic. At
some point we may combine the overlap and make nonhierarchical
validators operate directly on the AllConfigs, but that is a refactoring
for a later time. Where possible this CL reuses logic directly from
existing ValidationVisitors.

Bug: 137202024
Change-Id: I475006c4635166e01392056af7310410f22bb8e3