Enable importer to run as non-root user.
For git-sync, this required changing the security context from root user
to non-root user and also setting the filesystem group to the same
non-root user so that git-sync could still interact with the mounted
volumes.

For importer, this required setting the Docker image to run as a
non-root user and also creating a HOME directory that the non-root user
had permissions to write kube discovery cache within.

For resource quota admission controller, this required explicitly giving it the capability to bind port 443. Normally you can do this from the SecurityContext of the PodSpec but since k8s does not support ambient capabilities yet, I had to do this from within the Dockerfile instead.

Fixed: 129692589
Change-Id: Iee358bed3b64d26f878cea678b747869065f315e