Make status.From unnecessary in most cases
It is annoying to require status.From everywhere to convert from a
status.Error to status.MultiError.

This CL makes status.Error satisfy the status.MultiError interface since
by having it return the collection of itself. This removes boilerplate
in dozens of locations.

Minor change to util.go so it doesn't get stuck in an infinite loop as
it has conditional logic on satisfying the status.MultiError interface.

Change-Id: Iedf3d2aa158e5e297fe7c29709b4c88f9077946c