Restore namespace tombstoning to fix bugs.
We originally designed namespace tombstoning in go/delete-nsconfig-dd
and implemented for b/130737137. This was accidentally reverted during
the work for b/118385500. The tombstoning is reintroduced here.

We additionally designed a mechanism to avoid deleting all managed namespaces on syncer start up for b/112321191. This was accidentally reverted during the work for b/118189026. Rather than reintroducing the heavyweight "wait for importer" behavior, we are instead removing the syncer's fallback behavior of deleting a managed namespace when it has no matching namespaceconfig. This behavior is redundant with the tombstoning feature and leads to a dangerous edge case if a cluster previously had nomos installed and it later gets reinstalled and then managed namespaces may unexpectedly be deleted.

See linked bug below for more details on the customer issues this has
contributed to.

Fixed: 140199426
Change-Id: I4bdbae4e89d831709a607a362f5266d65aea6de6