Allow subdirs of cluster/ and clusterregistry/
Recursively parse these directories and don't throw errors when we
encounter objects in subdirectories.

Same for system/, because why not?

Really this only required deleting the code doing the check. We already
recursively parsed them. Surprisingly there weren't actually tests for
this, so I've added ones that verify the new behavior.

Bug: 141897848
Change-Id: Iab819be71edec031deb51c88735b00de067ec6c8