Increase timeouts in invalid policydir test
This test has been somewhat flaky lately, (roughly 20% failure rate
lately) and has always failed with a timeout waiting for the error to
appear. On the runs where the test failed, tests executing around the
same time took about 30-40% longer to execute (but still pass) than
they took in runs where the test in question did not fail. Since the
total execution time for this test in those cases is about 45 seconds,
a general slowdown of this magnitude would cause the timeouts we see.
Try increasing the timeout for the constiuent commands in this test and
see if reliability improves.

Bug: 142561691
Change-Id: I6c0d8b65a52a6224dc79341d7e24a3449d7e9c79