Actually use StrategicMergePatch
The current logic:
1) Creates a new schema with no registered schemas via
runtime.NewScheme()
2) Asks to see if the given type is registered via scheme.New() and
schema.IsUnversioned().
3) Since (2) trivially returns true, always tries a json merge patch.

We actually want to try a strategic merge patch whenever we can.

This CL instead checks the OpenAPI Schema provided by the APIServer.
Until sometime after 1.16, by default the APIServer ONLY returns the
base Kubernetes types in the OpenAPI schema, so the CRD schemas are
unavailable in a form we can pass to strategic merge patch.

Thus, this CL only fixes the issue for base Kubernetes types, not for
CRs.

To mitigate (in a different CL), we could download the relevant CRD and
convert it to the proto.Schema required. It isn't immediately clear how
to do this. However, this issue will resolve for CRDs either way in
1.17.

Bug: 142143147

Change-Id: I5dc2053272e222e4bd3f9aef9b4313ca3b8e36dd