Add check to nomos status for running on-cluster.
Nomos status performs a pre-check on all client connections to see if
they are valid and filter out ones that are not. This check does not
work if the nomos CLI binary is actually running on the cluster. We
check for a known environment variable to detect this and short circuit
the client check in that case.

Fixed: 140162660
Change-Id: I7bcf8f7c68da264fd9276a619859f39bc0606baf