Ensure metadata.annotation values are strings
The Kubernetes libraries silently ignore the entire metadata.annotations
field if any values are non-strings. This includes types that YAML
parsing automatically coerces to non-string values, like the unquoted
values `true` and `0000`.

This CL checks every value in metadata.annotations, then returns an
error pointing to the key with an invalud annotation value, suggesting
the user add quotes around it. This works because quoted `"true"` and
`"0000"` are parsed as strings.

As a follow-up we should file a bug against k8s.io/apimachinery since
the bug is somewhere in the codebase. It shouldn't silently discard
annotations if just one is invalid, ideally it should throw an error as
it does with invalid labels.

Bug: 143156032
Change-Id: I6d60dbbf1d618b84dcf4b9ca87f6d13d1da20a7b