Populate Namespaces in hydrate
This CL tackles two related issues:
1) We don't preserve the base path of Namespaces, so the source-path
annotation on Namespaces we sync is incorrect.
2) nomos hydrate doesn't write Namespaces

This CL modifies PathAnnotationVisitor to preserve the base path -
manually setting the annotation on the TreeNode to the path in the
Namespace. Obiously this assumes Namespace nodes have Namespaces, but
guaranteeing that suggests a refactoring far out of scope of this CL.

It also removes the source path annotation from AbstractNamespaces
because it's more trouble to maintain that behavior, and it isn't
actually used for anything.

This CL makes nomos hydrate copy the metadata from the NamespaceConfig.
The better solution would be to not discard the original Namespace, but
again, that's a much larger refactoring effort than the scope of this
CL.

Change-Id: I5dc13f85b564ddf1ac2c85611340b97c6bf2d53d