Trap during install in syllogi tests
The syllogi test traps only applied if the test failed after install. We
often see tests fail during install, so we would like the clean_up logs
then as well.

Also add a dev target to the syllogi buildfile for local dev on that
target without modifying the remote image, and increased logging in the
syllogi test.

Bug: 143682314
Change-Id: Iae5e5470bb8a56db3c3b41d29a6e5ed85233b828