Implement core.Object interface
We rely on implicit unsafe castings in many, many places of our code.

This CL vastly reduces the number of unsafe castings by implementing the
object.Object interface, which defines the minimal TypeMeta and
ObjectMeta fields we require. We cast to this type once after reading
objects from the disk.

We do have to do unsafe casting when copying objects because Go
doesn't allow generic interfaces, but as long as the types define their
own DeepCopyObject methods properly this will never be an issue.

This CL deletes a _lot_ of code that is now unnecessary as a result of
not having to do type casting everywhere. Most of this code change
takes advantage of that by having FileObject inherit our interface, we
don't have to jump through hoops to ensure we're dealing with the
correct objects. A lot of deleted code takes advantage of that we're no
longer using intermediate objects for the unsafe casts we were doing
previously.

In some cases methods became trivial since they didn't have to do type
casting, so I inlined the logic.

DisallowedFieldsValidator now safely checks for the existence of
optional metadata fields in the type before getting/validating the
values. This requires a lot of single-use interfaces, but ensures we're
compatible with all reasonable metadtata type definitions. The exception
is OwnerReference, which we check in a few locations.

I've included UID in core.Object since valid metadata definitions
MUST include the field, and we validate that the field is left blank.

In several cases I stumbled on libraries that used unsafe casting but
are unused. Rather than fix them, I have removed them.

Bug: 143557906

Change-Id: Iad0462e533561992993bbeb9e1c6319857a56bf1