Remove vet package and add missing error examples
New nomoserrors output: https://paste.googleplex.com/5922226207457280

Recall that the above tool is dev-only and in no way intended for
external use.

The vet package has long been a thorn in our side. Given an error code,
it's annoying to figure out where it came from since error codes are
decalred in a completely different location than where they're used.
Also, writing error messages is confusing as it's hard to keep the
context the error needs to communicate from a completely different file.

This CL moves all errors to the single location where they are used and
that doesn't cause an import loop.

This means we no longer have the god `vet` package from which all
packages inherit.

This CL also moves example error creation directly into nomoserrors. It
is much more convenient to be able to use the testing libraries than to
essentially recreate them, or to expose the error code to potentially
unstable test code.

In cases where there are multiple usage patterns for the same error
code, I've added all of the ones I could find as examples.

What's nice is that this has exposed glaring cases where error
documentation could be improved. Future developers will find it much
easier to explain the context of an error since they are closer to the
code that throws them.

Other than examples.go, the overwhelming majority of the code delta is
from moving code, deleting now-unnecessary files, and reducing import
complexity. In total this results in net-removal of 500 lines of code.

Change-Id: Ib1f76804fafd3fd9594e229b58a3ed33cff0178f