Improve KNV1027 and KNV1028
The main task here was to make KNV1027 and KNV1028's messages clearer.
Mainly, include more context on why the error was returned and provide
clearer courses of action to resolve the issue.

Old error messages: https://paste.googleplex.com/4610639597142016
New error messages: https://paste.googleplex.com/5727791125364736

Testing changes to error messages was annoying, so I added the --id
flag to nomoserrors so we can choose to only see error messages for a
particular KNV. This is useful when tuning and visually testing. It
treats --id=KNV1028 and --id=1028 as equivalent

The below refactorings were a result of that I was in the area:

Delete namespaceutil since util packages are a code smell.

Change IsReserved to IsControllerNamespace since right now that is all
it does. In the future if we decide to forbid other Namespace names -
which would be a probably-unacceptable backwards-incompatibility - we
can change it. I've left it as a function instead of replacing with the
trivial == check to make it more obvious which locations need to change
if any individual location must change.

Move isInvalid to directory_name_validator as that is the only location
which actually uses that method.

Include KNV1001 errors in KNV1028. It definitely falls under the
umbrella of "invalid namespace", but just needs its own error message.

Remove exhaustive tests of the Kubernetes libraries, e.g., it isn't our
job to unit test whether IsDNS1123Label works, just that *we* handle the
case where it reports/doesn't report an error correctly.

Bug: 144361025

Change-Id: I49b24012e2cd1897deca47df2c4c346798a8c963