Improve duplicate name errors
The message for KNV1002 is very bare-bones, and KNV1029 is misleading.
Also, the logic for detecting duplicates is both scattered in several
directories and has an odd interface.

This CL moves checking for duplicate Namespaces to the
duplicate_name_validator, which is a better place for it to live. It
also modifies the error messages to explicitly state why the error
happened and what the user needs to do to fix.

The biggest improvement here is the message for duplicate Namespaces.
Just saying "duplicate directories" is incomplete and doesn't get to the
*why* of the error. So KNV1002 has been merged into KNV1029.

Register KNV1001 and KNV1002 as deprecated.

Old messages: https://paste.googleplex.com/5689658929315840
New messages: https://paste.googleplex.com/4563759022473216

Bug: 144361025
Change-Id: Icbb085ea09b0a5c351ec406beb3653928ff3817a