Add intermediate flat format before AllConfigs
The issue we've had with applying code to both structured and
unstructured repos is that the logic never holds objects in the same
format. This CL adds a step to Parser.Parse that flattens the hierarchy
tree before reading it into the AllConfigs. Thus, both structured and
unstructured parsing now have an intermediate []FileObject that we can
run validation/selection/etc. on.

This change ends up requiring a few major changes, resulting in deleting
200 lines of special case code.

1. The logic formerly in RawParser.Parse that constructed the
AllConfigs is now generalized to work for both structured and
unstructured repos, and it is extracted out into its own method in
namespaceconfig/allconfigs.go.

2. For cleanliness, I've deleted the ImportToken and LoadTime from the
ast.Root and AllConfigs as they are no longer required. The values are
ONLY required when generating the AllConfigs, so carrying them around
everywhere just complicated interfaces.

3. As the *Config objects actually require metav1.Time, not time.Time,
I've changed the interfaces to ask for what they actually need. This
avoids the song and dance of different parts of the code having to do
the conversion since it is always the correct format.

4. New Flatten methods for ast.Root and ast.TreeNode that return the set
of objects they contained _that actually get materialized on the
cluster_. Thus, they ignore objects in AbstractNamespaces as by the
point Flatten() is called, they have already been copied down.

5. Namespaces are kept until the creation of the AllConfigs. On creating
the AllConfigs, they are consumed and turned into NamespaceConfigs.

6. testouptut.NewAllConfigs() that formats AllConfigs based on the input
FileObjects. There's no need for EVERY SINGLE TEST to respecify things
like Syncs, as that behavior is already sufficiently tested. While
terser, this resulted in hard-to-read long lines, so I've chopped down
all lines longer than 120 characters in parser_test.go for readability.

Change-Id: I5a49be6ced082552bc7044a347cdef4aad46ceee