Deduplicate hierarchical/nonhierarchical logic
This CL unifies a lot of logic, preferring functions that operate on
flat lists of objects since this is common to both hierarchical and
non-hierarchical parsing.

The overwhelming code change here is:
1) Deleting code/tests that were duplicated for both parsing types since
it now uses the nonhierarchical form
2) Copy+pasting code from hierarchical packages to nonhierarchical
packages
3) Rewriting hierarchical tests in the nonhierarchical form

This only actually changes already-existing logic in two locations:
1) Root.Flatten returns nil if Root is nil, in root.go
2) Parser.HydrateRoot does nonhierarchical validation, in parser.go

In a few cases I've renamed hierarchical validators to make it clearer
that they have hierarchy-specific logic to avoid confusion.

This does result in an unfortunate modification to parser.go where we
keep both the Root and the []FileObject, but that is temporary.

Change-Id: I7ad9bbf68ba89bd99aa7b975cb84dc31ebb93b69