Add tests for declaring CRD+CR in same CL
We weren't testing this before. As happens, the test infrastructure
didn't allow for this sort of test, so I've modified it to do so.

Specifically, the TestClient didn't properly forward the extra resources
to the Kubernetes parsing library.

Remove the hack that put the Engineer resource in TestDynamicResources
since it is only required for a few tests. This has the advantage of
ensuring we handle resource lists consistently when parsing.

Change-Id: I968e1d62d98a1165b8a44182c8d9161e1c0ca60f