Define non-hierarchical ClusterSelector logic
We want to allow ClusterSelectors in non-hierarchical mode, so we need
the ClusterSelector logic to work on both hierarchical and
non-hierarchical repos. As turns out, operating on []ast.FileObject is
much simpler than doing it hierarchically.

The overwhelming majority of this CL is deleting unnecessary code. Begin
reviewing in cluster_selectors.go.

Before we did a complex operation where we inlined ClusterSelectors
inside the objects referencing them as JSON, then parsed the JSON into a
LabelSelector before using it.

Now we do everything in one place, resulting in about 800 fewer lines of
code. The main files to pay attention to are cluster_selectors.go, which
holds the full ClusterSelector logic, and root.go that now maintains the
objects in clusterregistry/ when flattening to []ast.FileObject.

This CL doesn't touch NamespaceSelectors for the obvious reason that the
CL would be even larger. Also obviously, there should be a follow-up
that does this to make the patterns consistent.

Change-Id: I46518d09e53d8bb46208bf930490772f2b812682