Refactor APIInfo
Define APIInfo as a map with methods defined on it. This is more
idiomatic Go and makes tests including APIInfos less annoying.

Refactor NewAPIInfo to return all parsing errors rather than just the
first.

Greatly expand comment explaining what APIInfo is _for_, particularly
that it is for determining what to sync to the cluster. The point is to
get the intersection between the types we want to sync and the types
currently available to the server, as we don't want to sync things like
CRs that aren't yet uploaded.

Change-Id: I287eeb9f91c2e8f783a33fed07cb10369e5e6fe2