Check for selector uniqueness
We currently allow users to have ClusterSelectors and NamespaceSelectors
with duplicate names, as we remove those types before checking for
duplicate objects. This can lead to ambiguous scenarios where it isn't
clear which selector applies since multiple have the same label.

This CL implement checking *Selector uniqueness before running them.

We can't do this when we check for duplicate objects, as then it is too
late, as we don't check for duplicate objects until AFTER we apply
*Selection.

We must do ClusterSelector uniqueness checking BEFORE ClusterSelection,
and NamespaceSelector uniquenss AFTER ClusterSelection because
NamespaceSelectors may be cluster-selected.

This implements the logic for detecting duplicate NamespaceSelectors, as
well as the presumed future code change in parser.go commented out. That
comment will be uncommented once non-hierarchical NamespaceSelector logic
is implemented.

Change-Id: I04d59a6c989cc7f1b25ee38f4f7b946df6b9973e