Consistently read APIResources from server
We aren't consistent in how we read resources from the APIServer,
resulting in slightly different behaviors and error messages. Also, in
several cases we panic if the APIServer returns an error, rather than
gracefully returning an error. This CL fixes that.

First, delete api_resources_lister since that is completely unused and
is getting the way of refactoring.

Move ValidateInstallation to controller.go since that is where it is
actually used.

Define consistent means of getting server resources from a
DiscoveryClient. As-is we're returning different error messages
depending on where it failed, when it's really the same error.

Use Scoper to test for installation.

Delete less_versions since that is also unused.

Change-Id: Iea255f07acc99a063f0a05086408ddb2af309385