Remove apiinfo from Root.Extensions
Finally remove APIInfo from Root.Extensions, ensuring at compile-time
that the Scoper exists rather than relying on things to be called in a
particular order at runtime.

This lets us get rid of a lot of code as the APIInfoVisitor no longer
has a reason to exist, and makes testing easier since we don't have to
special-case tests with an APIInfo stored in the root.

Rename KnownResourceValidator to HierarchyConfigScopeValidator, which is
more descriptive of what it is actually doing, and makes it obvious from
the name that this is specific to hierarchy enabled. Rewrite to operate
on list of FileObjects so passing in the scoper argument is possible.

Rename Scope to TopLevelDirectoryValidator, as above to be more
descriptive. As above, rewrite to operate on a flat list of FileObjects
to avoid the Root.Extensions messiness.

Add Parser.getScoper convenience method to avoid repeating those lines
of logic.

In fake.go, define a standard extensible Scoper for use in testing. this
avoid having to manually create ResourceLists just to test scoping logic
for parser_test.

Rename "Process" to "GetCRDs" since that makes it much more obvious what
the method is doing.

Yes, we do duplicate the effort of creating the Scoper twice, but that
is temporary.

Change-Id: I240b400e3de9dab3be1faed32938a87c3b8c832b