Upgrade k8s dependencies to 1.16.3
Our k8s dependencies are currently pinned to 1.12.3
This is starting to be a bit archaic, and google3 is on quite a recent
set of k8s third-party dependencies, so it makes sense to try to keep
this up-to-date. We chose 1.16.3 as the most recent stable release
available.

Re-ran client-gen via scripts/generate-clientset.sh with the new 1.16.3
code generator, but this mostly just added more license files.

Other code changes made in non-vendor are to address API changes. Mock
changes were done manually because I had some issues getting mockgen to
work.

Change-Id: I78a6326815614a66a789a35eea63af3d71d6f0c8