Fix scope checking for generated ResourceQuotas
When a user defines a ResourceQuota with hierarchicalQuota enabled, we
generate a bunch of *new* ResourceQuotas to sync to the server. However,
we do not assign them filepaths.

My previous change moved the check that ensures objects are located in
the correct top-level directories to *after* we do this generation.
Thus, validation now fails if a user uses HierarchicalQuota.

We did not formerly have unit test coverage of this case, so it didn't
fail until e2e testing.

This CL modifies scope checking to specifically look for these generated
ResourceQuotas, and adds unit tests that verify that we handle this case
correctly.

Also fix previously-incorrect wording for such messages.

Change-Id: I45249ea94aae873b74428635e1e11f188f65295d