Fix APIService interaction.
This fixes a few issues in importer / syncer with regard to APIService
changes on the cluster.  First, this ensures that the importer only
caches the directory state on success so that it will always eventually
retry if there was a failure to update any resource.  Next, this fixes
an error for APIDiscovery where the discovery would return an error if
any of the calls failed despite some succeding.  We can (maybe) safely
ignore that specific type of error and continue going forward since
that's better than the infra getting stuck.

Change-Id: Ieca2bd9e479005bbbd5a84d2d9f9850d1d8ca3b0