Fix varargs convention on CRD list.
We have a bunch of places where we provide the CRD list as varargs
despite already having it in a slice.  Varargs are useful if a
programmer will be manually providing a variable number of args, but in
our case we are only providing a single slice, so varargs aren't the
right fit for the situation.  For handling APIService, we have to
provide additional args to handle resources that the aggregated server
will expose.  The varargs get in the way, so this is a required refactor
before that work begins.

Bug: 144397545
Change-Id: I6a9e0402a91acd9fd364a235aaca7f97bca4853f