Refactor cluster_selectors.go
The upcoming namespace_selector non-hierarchical implementation requires
a doubly-nested map of both Namespace names and Selector names. This
gets confusing to read, and makes it hard to remember which string is at
which level of nesting in the map.

This CL preemptively creates selectorName and namespaceName string
aliases to ensure at compile-time that we can't make mistakes.

So basically this CL's purpose is to reduce the size of an upcoming CL
and make the meaning of that CL's code clearer.

Change-Id: I451efc362e3c9f1e14bb816f89e9f3bc6c4ff5cd