Collapse top level directory validation
ACM requires that users declare objects in specific subdirectories of
the root directory. We do this check in three different locations of the
code, resulting in fragmented logic.

This CL collapses all of these checks into a single validation check,
TopLevelDirectoryValidator.

Related - ObjectScope is now a bool alias, IsNamespaced. This means when
we check for the scope of a type, we no longer have to switch, but can
use an if statement. This gets rid of a lot of code checking for
conditions we know can't happen, i.e. for ObjectScope to be anything but
NamespaceScoped or ClusterScoped.

Change-Id: Ie3725a3ab104c654ca5965c1026990035cc55bfe