Add SelectorAnnotationValidator
Right now we validate the namespace-selector and cluster-selector
annotation in several different places in the code. We actually don't
verify that Cluster-scoped objects *don't* declare the
namespace-selector annotation.

This CL implements non-hierarchical *-selector annotation validation.
While longer than the current code, this allows the final
nonhierarchical NamespaceSelector CL to be shorter. As this validation
is now in one place, we have exactly one set of tests declaring which
types of objects are allowed/not allowed to have each *-selector
annotation.

This CL also refactors TreeNode.flatten() so that the new validator
actually sees NamespaceSelectors.

Change-Id: If67aa4876c863824a2e6b568dc21d914e81747fb