Simplify parser creation, usage, and testing
We've recently removed a lot of parameters required for Parser to
function, and since Bexit is complete we no longer need to support
multiple lines of parsing.

Most of this change is just the fact that parser.Parse is moved to
earlier in its file.

This CL removes NomosParserConfig since all it holds now is the root
path to the Nomos repository.

This CL also makes NewParser accept a filesystem.Reader, so we can pass
in a test reader rather than requiring it to read from a file system (we
can do this because of the new changes to how we read objects).

This CL also un-exports a lot of Parser methods that were previously
needed for testing.

Doing this required defining RootedPath, an interface for paths that are
either a repository root or a path relative to one, specifically for use
by the filesystem.Reader functions. It makes calling .Read() less of a
hassle.

Change-Id: I546957ecf7960549f54e9c50a2ae2bd2f7de1bdc