Add --no-api-server-check to nomos vet
A big use case for nomos vet is being able to use it in CI to check that
configuration is valid before pushing to a master branch.
go/serverless-nomos-vet

This CL adds --no-api-server-check, which only disables checks that
would require an API server. The only errors this does not catch
are when the user has a non-standard type defined on the APIServer by
either an unmanaged CRD or an API extension server.

Thus, in cases where we would have thrown an error before about not
recognizing the type, we instead throw no error. This does not prevent
other checks from being done on those objects, such as invalid
annotations or metadata.name collisions.

Fixes: 146063950

Change-Id: I3424632c9166e1954f5e1778e8366f81c8f46c18