Simplify using Parser.Parse()
The importToken and loadTime arguments are only used for one thing -
annotating an AllConfigs at creation. This complicates using
Parser.Parse(), resulting in a lot of boilerplate.

Similarly, NewAllConfigs did require a Scoper, but really shouldn't. We
By the time we pass objects to NewAllConfigs, we've already assigned
namespace-scoped objects metadata.namespace, so we can just check that
value rather than requiring generating a second scoper just for that.
(Remember that we've already done scope-checking at this point.)

This makes calling Parser.Parse and reasoning about it much easier, as
there are fewer moving parts.

Change-Id: I0591f07f17d796580691e529e3e68828db7c5d5c